data/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
click==8.1.3
    # via typer
colorama==0.4.6
    # via typer
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
psycopg2-binary==2.9.6
    # via -r requirements.in
pydantic==1.10.9
    # via -r requirements.in
pygments==2.15.1
    # via rich
rich==13.4.2
    # via typer
shellingham==1.5.0.post1
    # via typer
typer[all]==0.9.0
    # via -r requirements.in
typing-extensions==4.6.3
    # via
    #   pydantic
    #   typer
